# specs/001-qr-platform/contracts/api.yaml
openapi: 3.0.0
info:
  title: QR Platform API
  version: 1.0.0
  description: RESTful API for QR code platform
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /scan/{id}:
    get:
      summary: Scan QR code (redirect or 404)
      tags: [QR Codes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to target URL or image
          headers:
            Location:
              type: string
        '404':
          description: QR code not found, paused, or archived

  # Admin
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...